lib: lte_link_control: refactor into modules
Divides the LTE Link Controller into modules.
