+++ applications/asset_tracker_v2/prj.conf
@@ -46,6 +46,7 @@ CONFIG_LTE_PSM_REQ=y
 CONFIG_LTE_PSM_REQ_RPTAU="11000001"
 ### 20 seconds active time.
 CONFIG_LTE_PSM_REQ_RAT="00001010"
+CONFIG_LTE_LC_NCELLMEAS=y
 
 # Settings - Used to store real-time device configuration to flash.
 CONFIG_SETTINGS=y
