+++ include/modem/lte_lc.h
@@ -214,12 +214,14 @@ enum lte_lc_evt_type {
 	 */
 	LTE_LC_EVT_NW_REG_STATUS,
 
+#if defined(CONFIG_LTE_LC_PSM) || defined(__DOXYGEN__)
 	/**
 	 * PSM parameters provided by the network.
 	 *
 	 * The associated payload is the @c psm_cfg member of type @ref lte_lc_psm_cfg in the event.
 	 */
 	LTE_LC_EVT_PSM_UPDATE,
+#endif
 
 	/**
 	 * eDRX parameters provided by the network.
@@ -324,80 +326,6 @@ enum lte_lc_rrc_mode {
 	LTE_LC_RRC_MODE_CONNECTED	= 1,
 };
 
-/** Power Saving Mode (PSM) configuration. */
-struct lte_lc_psm_cfg {
-	/** Periodic Tracking Area Update interval in seconds. */
-	int tau;
-
-	/** Active-time (time from RRC idle to PSM) in seconds or @c -1 if PSM is deactivated. */
-	int active_time;
-};
-
-/** eDRX configuration. */
-struct lte_lc_edrx_cfg {
-	/**
-	 * LTE mode for which the configuration is valid.
-	 *
-	 * If the mode is @ref LTE_LC_LTE_MODE_NONE, access technology is not using eDRX.
-	 */
-	enum lte_lc_lte_mode mode;
-
-	/** eDRX interval in seconds. */
-	float edrx;
-
-	/** Paging time window in seconds. */
-	float ptw;
-};
-
-/** Maximum timing advance value. */
-#define LTE_LC_CELL_TIMING_ADVANCE_MAX		20512
-/** Invalid timing advance value. */
-#define LTE_LC_CELL_TIMING_ADVANCE_INVALID	65535
-/** Maximum EARFCN value. */
-#define LTE_LC_CELL_EARFCN_MAX			262143
-/** Unknown or undetectable RSRP value. */
-#define LTE_LC_CELL_RSRP_INVALID		255
-/** Unknown or undetectable RSRQ value. */
-#define LTE_LC_CELL_RSRQ_INVALID		255
-/** Cell ID value not valid. */
-#define LTE_LC_CELL_EUTRAN_ID_INVALID		UINT32_MAX
-/** Maximum cell ID value. */
-#define LTE_LC_CELL_EUTRAN_ID_MAX		268435455
-/** Tracking Area Code value not valid. */
-#define LTE_LC_CELL_TAC_INVALID			UINT32_MAX
-/** Time difference not valid. */
-#define LTE_LC_CELL_TIME_DIFF_INVALID		0
-
-/** Neighbor cell information. */
-struct lte_lc_ncell {
-	/** EARFCN per 3GPP TS 36.101. */
-	uint32_t earfcn;
-
-	/**
-	 * Difference of current cell and neighbor cell measurement, in the range
-	 * -99999 ms < time_diff < 99999 ms. @ref LTE_LC_CELL_TIME_DIFF_INVALID if the value is not
-	 * valid.
-	 */
-	int time_diff;
-
-	/** Physical cell ID. */
-	uint16_t phys_cell_id;
-
-	/**
-	 * RSRP.
-	 *
-	 * Format is the same as for @c rsrp member of struct @ref lte_lc_cell.
-	 */
-	int16_t rsrp;
-
-	/**
-	 * RSRQ.
-	 *
-	 * Format is the same as for @c rsrq member of struct @ref lte_lc_cell.
-	 */
-	int16_t rsrq;
-};
-
 /** Cell information. */
 struct lte_lc_cell {
 	/** Mobile Country Code. */
@@ -495,6 +423,70 @@ struct lte_lc_cell {
 	int16_t rsrq;
 };
 
+/** eDRX configuration. */
+struct lte_lc_edrx_cfg {
+	/**
+	 * LTE mode for which the configuration is valid.
+	 *
+	 * If the mode is @ref LTE_LC_LTE_MODE_NONE, access technology is not using eDRX.
+	 */
+	enum lte_lc_lte_mode mode;
+
+	/** eDRX interval in seconds. */
+	float edrx;
+
+	/** Paging time window in seconds. */
+	float ptw;
+};
+
+/** Maximum timing advance value. */
+#define LTE_LC_CELL_TIMING_ADVANCE_MAX		20512
+/** Invalid timing advance value. */
+#define LTE_LC_CELL_TIMING_ADVANCE_INVALID	65535
+/** Maximum EARFCN value. */
+#define LTE_LC_CELL_EARFCN_MAX			262143
+/** Unknown or undetectable RSRP value. */
+#define LTE_LC_CELL_RSRP_INVALID		255
+/** Unknown or undetectable RSRQ value. */
+#define LTE_LC_CELL_RSRQ_INVALID		255
+/** Cell ID value not valid. */
+#define LTE_LC_CELL_EUTRAN_ID_INVALID		UINT32_MAX
+/** Maximum cell ID value. */
+#define LTE_LC_CELL_EUTRAN_ID_MAX		268435455
+/** Tracking Area Code value not valid. */
+#define LTE_LC_CELL_TAC_INVALID			UINT32_MAX
+/** Time difference not valid. */
+#define LTE_LC_CELL_TIME_DIFF_INVALID		0
+
+/** Neighbor cell information. */
+struct lte_lc_ncell {
+	/** EARFCN per 3GPP TS 36.101. */
+	uint32_t earfcn;
+
+	/**
+	 * Difference of current cell and neighbor cell measurement, in the range
+	 * -99999 ms < time_diff < 99999 ms. @ref LTE_LC_CELL_TIME_DIFF_INVALID if the value is not
+	 * valid.
+	 */
+	int time_diff;
+
+	/** Physical cell ID. */
+	uint16_t phys_cell_id;
+
+	/**
+	 * RSRP.
+	 *
+	 * Format is the same as for @c rsrp member of struct @ref lte_lc_cell.
+	 */
+	int16_t rsrp;
+
+	/**
+	 * RSRQ.
+	 *
+	 * Format is the same as for @c rsrq member of struct @ref lte_lc_cell.
+	 */
+	int16_t rsrq;
+};
 
 /** Results of neighbor cell measurements. */
 struct lte_lc_cells_info {
@@ -524,38 +516,99 @@ struct lte_lc_cells_info {
 	struct lte_lc_cell *gci_cells;
 };
 
-/** Modem sleep type. */
-enum lte_lc_modem_sleep_type {
-	/** Power Saving Mode (PSM). */
-	LTE_LC_MODEM_SLEEP_PSM			= 1,
+#if defined(CONFIG_LTE_LC_XMODEMSLEEP) || defined(__DOXYGEN__)
+#include "lte_lc/lte_lc_xmodemsleep.h"
+#endif
 
-	/** RF inactivity, for example eDRX. */
-	LTE_LC_MODEM_SLEEP_RF_INACTIVITY	= 2,
+#if defined(CONFIG_LTE_LC_REDMOB) || defined(__DOXYGEN__)
+#include "lte_lc/lte_lc_redmob.h"
+#endif
 
-	/** Limited service or out of coverage. */
-	LTE_LC_MODEM_SLEEP_LIMITED_SERVICE	= 3,
+#if defined(CONFIG_LTE_LC_PSM) || defined(__DOXYGEN__)
+#include "lte_lc/lte_lc_psm.h"
+#endif
 
-	/** Flight mode. */
-	LTE_LC_MODEM_SLEEP_FLIGHT_MODE		= 4,
+#if defined(CONFIG_LTE_LC_PERIODICSEARCHCONF) || defined(__DOXYGEN__)
+#include "lte_lc/lte_lc_periodicsearchconf.h"
+#endif
 
+/** Modem domain events. */
+enum lte_lc_modem_evt {
 	/**
-	 * Proprietary PSM.
+	 * Indicates that a light search has been performed.
 	 *
-	 * @note This type is only supported by modem firmware versions >= 2.0.0.
+	 * This event gives the application the chance to stop using more power when searching
+	 * networks in possibly weaker radio conditions. Before sending this event, the modem
+	 * searches for cells based on previous cell history, measures the radio conditions, and
+	 * makes assumptions on where networks might be deployed. This event means that the modem
+	 * has not found a network that it could select based on the 3GPP network selection rules
+	 * from those locations. It does not mean that there are no networks to be found in the
+	 * area. The modem continues more thorough searches automatically after sending this status.
 	 */
-	LTE_LC_MODEM_SLEEP_PROPRIETARY_PSM	= 7,
-};
-
-/** Modem sleep information. */
-struct lte_lc_modem_sleep {
-	/** Sleep type. */
-	enum lte_lc_modem_sleep_type type;
+	LTE_LC_MODEM_EVT_LIGHT_SEARCH_DONE,
 
 	/**
-	 * Sleep time in milliseconds. If this value is set to -1, the sleep is considered infinite.
+	 * Indicates that a network search has been performed.
+	 *
+	 * The modem has found a network that it can select according to the 3GPP network selection
+	 * rules or all frequencies have been scanned and a suitable cell was not found. In the
+	 * latter case, the modem enters normal limited service state functionality and performs
+	 * scan for service periodically.
 	 */
-	int64_t time;
+	LTE_LC_MODEM_EVT_SEARCH_DONE,
+
+	/**
+	 * The modem has detected a reset loop.
+	 *
+	 * A reset loop indicates that the modem restricts network attach attempts for the next 30
+	 * minutes. The timer does not run when the modem has no power or while it stays in the
+	 * reset loop. The modem counts all the resets where the modem is not gracefully
+	 * deinitialized with `AT+CFUN=0` or using an API performing the equivalent operation, such
+	 * as setting the functional mode to @ref LTE_LC_FUNC_MODE_POWER_OFF.
+	 */
+	LTE_LC_MODEM_EVT_RESET_LOOP,
+
+	/** Battery voltage is low and the modem is therefore deactivated. */
+	LTE_LC_MODEM_EVT_BATTERY_LOW,
+
+	/** The device is overheated and the modem is therefore deactivated. */
+	LTE_LC_MODEM_EVT_OVERHEATED,
+
+	/** The modem does not have an IMEI. */
+	LTE_LC_MODEM_EVT_NO_IMEI,
+
+	/**
+	 * Selected CE level in RACH procedure is 0, see 3GPP TS 36.331 for details.
+	 *
+	 * @note This event is only supported by modem firmware versions >= 2.0.0.
+	 */
+	LTE_LC_MODEM_EVT_CE_LEVEL_0,
+
+	/**
+	 * Selected CE level in RACH procedure is 1, see 3GPP TS 36.331 for details.
+	 *
+	 * @note This event is only supported by modem firmware versions >= 2.0.0.
+	 */
+	LTE_LC_MODEM_EVT_CE_LEVEL_1,
+
+	/**
+	 * Selected CE level in RACH procedure is 2, see 3GPP TS 36.331 for details.
+	 *
+	 * @note This event is only supported by modem firmware versions >= 2.0.0.
+	 */
+	LTE_LC_MODEM_EVT_CE_LEVEL_2,
+
+	/**
+	 * Selected CE level in RACH procedure is 3, see 3GPP TS 36.331 for details.
+	 *
+	 * @note This event is only supported by modem firmware versions >= 2.0.0.
+	 */
+	LTE_LC_MODEM_EVT_CE_LEVEL_3,
 };
+
+#if defined(CONFIG_LTE_LC_XFACTORYRESET) || defined(__DOXYGEN__)
+#include "lte_lc/lte_lc_xfactoryreset.h"
+#endif
 
 /** Energy consumption estimate. */
 enum lte_lc_energy_estimate {
@@ -633,105 +686,6 @@ enum lte_lc_ce_level {
 
 	/** Not known. */
 	LTE_LC_CE_LEVEL_UNKNOWN	= UINT8_MAX,
-};
-
-/** Reduced mobility mode. */
-enum lte_lc_reduced_mobility_mode {
-	/** Functionality according to the 3GPP relaxed monitoring feature. */
-	LTE_LC_REDUCED_MOBILITY_DEFAULT = 0,
-
-	/** Enable Nordic-proprietary reduced mobility feature. */
-	LTE_LC_REDUCED_MOBILITY_NORDIC = 1,
-
-	/**
-	 * Full measurements for best possible mobility.
-	 *
-	 * Disable the 3GPP relaxed monitoring and Nordic-proprietary reduced mobility features.
-	 */
-	LTE_LC_REDUCED_MOBILITY_DISABLED = 2,
-};
-
-/** Modem domain events. */
-enum lte_lc_modem_evt {
-	/**
-	 * Indicates that a light search has been performed.
-	 *
-	 * This event gives the application the chance to stop using more power when searching
-	 * networks in possibly weaker radio conditions. Before sending this event, the modem
-	 * searches for cells based on previous cell history, measures the radio conditions, and
-	 * makes assumptions on where networks might be deployed. This event means that the modem
-	 * has not found a network that it could select based on the 3GPP network selection rules
-	 * from those locations. It does not mean that there are no networks to be found in the
-	 * area. The modem continues more thorough searches automatically after sending this status.
-	 */
-	LTE_LC_MODEM_EVT_LIGHT_SEARCH_DONE,
-
-	/**
-	 * Indicates that a network search has been performed.
-	 *
-	 * The modem has found a network that it can select according to the 3GPP network selection
-	 * rules or all frequencies have been scanned and a suitable cell was not found. In the
-	 * latter case, the modem enters normal limited service state functionality and performs
-	 * scan for service periodically.
-	 */
-	LTE_LC_MODEM_EVT_SEARCH_DONE,
-
-	/**
-	 * The modem has detected a reset loop.
-	 *
-	 * A reset loop indicates that the modem restricts network attach attempts for the next 30
-	 * minutes. The timer does not run when the modem has no power or while it stays in the
-	 * reset loop. The modem counts all the resets where the modem is not gracefully
-	 * deinitialized with `AT+CFUN=0` or using an API performing the equivalent operation, such
-	 * as setting the functional mode to @ref LTE_LC_FUNC_MODE_POWER_OFF.
-	 */
-	LTE_LC_MODEM_EVT_RESET_LOOP,
-
-	/** Battery voltage is low and the modem is therefore deactivated. */
-	LTE_LC_MODEM_EVT_BATTERY_LOW,
-
-	/** The device is overheated and the modem is therefore deactivated. */
-	LTE_LC_MODEM_EVT_OVERHEATED,
-
-	/** The modem does not have an IMEI. */
-	LTE_LC_MODEM_EVT_NO_IMEI,
-
-	/**
-	 * Selected CE level in RACH procedure is 0, see 3GPP TS 36.331 for details.
-	 *
-	 * @note This event is only supported by modem firmware versions >= 2.0.0.
-	 */
-	LTE_LC_MODEM_EVT_CE_LEVEL_0,
-
-	/**
-	 * Selected CE level in RACH procedure is 1, see 3GPP TS 36.331 for details.
-	 *
-	 * @note This event is only supported by modem firmware versions >= 2.0.0.
-	 */
-	LTE_LC_MODEM_EVT_CE_LEVEL_1,
-
-	/**
-	 * Selected CE level in RACH procedure is 2, see 3GPP TS 36.331 for details.
-	 *
-	 * @note This event is only supported by modem firmware versions >= 2.0.0.
-	 */
-	LTE_LC_MODEM_EVT_CE_LEVEL_2,
-
-	/**
-	 * Selected CE level in RACH procedure is 3, see 3GPP TS 36.331 for details.
-	 *
-	 * @note This event is only supported by modem firmware versions >= 2.0.0.
-	 */
-	LTE_LC_MODEM_EVT_CE_LEVEL_3,
-};
-
-/** Type of factory reset to perform. */
-enum lte_lc_factory_reset_type {
-	/** Reset all modem data to factory settings. */
-	LTE_LC_FACTORY_RESET_ALL = 0,
-
-	/** Reset user-configurable data to factory settings. */
-	LTE_LC_FACTORY_RESET_USER = 1,
 };
 
 /**
@@ -977,168 +931,6 @@ struct lte_lc_ncellmeas_params {
 	uint8_t gci_count;
 };
 
-
-/** Search pattern type. */
-enum lte_lc_periodic_search_pattern_type {
-	/** Range search pattern. */
-	LTE_LC_PERIODIC_SEARCH_PATTERN_RANGE = 0,
-
-	/** Table search pattern. */
-	LTE_LC_PERIODIC_SEARCH_PATTERN_TABLE = 1,
-};
-
-/** Configuration for periodic search of type @ref LTE_LC_PERIODIC_SEARCH_PATTERN_RANGE. */
-struct lte_lc_periodic_search_range_cfg {
-	/**
-	 * Sleep time between searches in the beginning of the range.
-	 *
-	 * Allowed values: 0 - 65535 seconds.
-	 */
-	uint16_t initial_sleep;
-
-	/**
-	 * Sleep time between searches in the end of the range.
-	 *
-	 * Allowed values: 0 - 65535 seconds.
-	 */
-	uint16_t final_sleep;
-
-	/**
-	 * Optional target time in minutes for achieving the @c final_sleep value.
-	 *
-	 * This can be used to determine angle factor between the initial and final sleep times.
-	 * The timeline for the @c time_to_final_sleep starts from the beginning of the search
-	 * pattern. If given, the value cannot be greater than the value of the
-	 * @c pattern_end_point  value in the same search pattern. If not given, the angle factor
-	 * is calculated by using the @c pattern_end_point value so, that the @c final_sleep
-	 * value is reached at the point of @c pattern_end_point.
-	 *
-	 * Allowed values:
-	 * * -1: Not used
-	 * * 0 - 1080 minutes
-	 */
-	int16_t time_to_final_sleep;
-
-	/**
-	 * Time that must elapse before entering the next search pattern.
-	 *
-	 * The timeline for @c pattern_end_point starts from the beginning of the limited service
-	 * starting point, which is the moment when the first sleep period started.
-	 *
-	 * Allowed values:
-	 * * 0 - 1080 minutes.
-	 */
-	int16_t pattern_end_point;
-};
-
-/**
- * Configuration for periodic search of type @ref LTE_LC_PERIODIC_SEARCH_PATTERN_TABLE.
- *
- * 1 to 5 sleep time values for sleep between searches can be configured. It is mandatory to provide
- * @c val_1, while the rest are optional. Unused values must be set to -1. After going through all
- * values, the last value of the last search pattern is repeated, if not configured differently by
- * the @c loop or @c return_to_pattern parameters.
- *
- * Allowed values:
- * * -1: Value unused.
- * * 0 - 65535 seconds.
- */
-struct lte_lc_periodic_search_table_cfg {
-	/** Mandatory sleep time. */
-	int val_1;
-
-	/**
-	 * Optional sleep time.
-	 *
-	 * Must be set to -1 if not used.
-	 */
-	int val_2;
-
-	/**
-	 * Optional sleep time.
-	 *
-	 * @c val_2 must be configured for this parameter to take effect.
-	 *
-	 *  Must be set to -1 if not used.
-	 */
-	int val_3;
-
-	/**
-	 * Optional sleep time.
-	 *
-	 * @c val_3 must be configured for this parameter to take effect.
-	 *
-	 * Must be set to -1 if not used.
-	 */
-	int val_4;
-
-	/**
-	 * Optional sleep time.
-	 *
-	 * @c val_4 must be configured for this parameter to take effect.
-	 *
-	 * Must be set to -1 if not used.
-	 */
-	int val_5;
-};
-
-/**
- * Periodic search pattern.
- *
- * A search pattern may be of either 'range' or 'table' type.
- */
-struct lte_lc_periodic_search_pattern {
-	/** Search pattern type. */
-	enum lte_lc_periodic_search_pattern_type type;
-
-	union {
-		/** Configuration for periodic search of type 'range'. */
-		struct lte_lc_periodic_search_range_cfg range;
-
-		/** Configuration for periodic search of type 'table'. */
-		struct lte_lc_periodic_search_table_cfg table;
-	};
-};
-
-/** Periodic search configuration. */
-struct lte_lc_periodic_search_cfg {
-	/**
-	 * Indicates if the last given pattern is looped from the beginning when the pattern has
-	 * ended.
-	 *
-	 * If several patterns are configured, this impacts only the last pattern.
-	 */
-	bool loop;
-
-	/**
-	 * Indicates if the modem can return to a given search pattern with shorter sleeps, for
-	 * example, when radio conditions change and the given pattern index has already been
-	 * exceeded.
-	 *
-	 * Allowed values:
-	 * * 0: No return pattern.
-	 * * 1 - 4: Return to search pattern index 1..4.
-	 */
-	uint16_t return_to_pattern;
-
-	/**
-	 * Indicates if band optimization shall be used.
-	 *
-	 * * 0: No optimization. Every periodic search shall be all band search.
-	 * * 1: Use default optimizations predefined by modem. Predefinition depends on
-	 *      the active data profile.
-	 *      See "nRF91 AT Commands - Command Reference Guide" for more information.
-	 * * 2 - 20: Every n periodic search must be an all band search.
-	 */
-	uint16_t band_optimization;
-
-	/** The number of valid patterns. Range 1 - 4. */
-	size_t pattern_count;
-
-	/** Array of periodic search patterns. */
-	struct lte_lc_periodic_search_pattern patterns[4];
-};
-
 /** Callback for modem functional mode changes. */
 struct lte_lc_cfun_cb {
 	void (*callback)(enum lte_lc_func_mode, void *ctx);
@@ -1171,8 +963,10 @@ struct lte_lc_evt {
 		/** Payload for event @ref LTE_LC_EVT_RRC_UPDATE. */
 		enum lte_lc_rrc_mode rrc_mode;
 
+#if defined(CONFIG_LTE_LC_PSM) || defined(__DOXYGEN__)
 		/** Payload for event @ref LTE_LC_EVT_PSM_UPDATE. */
 		struct lte_lc_psm_cfg psm_cfg;
+#endif
 
 		/** Payload for event @ref LTE_LC_EVT_EDRX_UPDATE. */
 		struct lte_lc_edrx_cfg edrx_cfg;
@@ -1183,11 +977,13 @@ struct lte_lc_evt {
 		/** Payload for event @ref LTE_LC_EVT_LTE_MODE_UPDATE. */
 		enum lte_lc_lte_mode lte_mode;
 
+#if defined(CONFIG_LTE_LC_XMODEMSLEEP) || defined(__DOXYGEN__)
 		/**
 		 * Payload for events @ref LTE_LC_EVT_MODEM_SLEEP_EXIT_PRE_WARNING,
 		 * @ref LTE_LC_EVT_MODEM_SLEEP_EXIT and @ref LTE_LC_EVT_MODEM_SLEEP_ENTER.
 		 */
 		struct lte_lc_modem_sleep modem_sleep;
+#endif
 
 		/** Payload for event @ref LTE_LC_EVT_MODEM_EVENT. */
 		enum lte_lc_modem_evt modem_evt;
@@ -1663,96 +1459,6 @@ int lte_lc_modem_events_enable(void);
  * @retval -EFAULT if AT command failed.
  */
 int lte_lc_modem_events_disable(void);
-
-/**
- * Configure periodic searches.
- *
- * This configuration affects the periodic searches that the modem performs in limited service state
- * to obtain normal service. See @ref lte_lc_periodic_search_cfg and
- * "nRF91 AT Commands - Command Reference Guide" for more information and in-depth explanations of
- * periodic search configuration.
- *
- * @param[in] cfg Periodic search configuration.
- *
- * @retval 0 if the configuration was successfully sent to the modem.
- * @retval -EINVAL if an input parameter was @c NULL or contained an invalid pattern count.
- * @retval -EFAULT if an AT command could not be sent to the modem.
- * @retval -EBADMSG if the modem responded with an error to an AT command.
- */
-int lte_lc_periodic_search_set(const struct lte_lc_periodic_search_cfg *const cfg);
-
-/**
- * Get the configured periodic search parameters.
- *
- * @param[out] cfg Periodic search configuration.
- *
- * @retval 0 if a configuration was found and populated to the provided pointer.
- * @retval -EINVAL if input parameter was @c NULL.
- * @retval -ENOENT if periodic search configuration was not set.
- * @retval -EFAULT if an AT command failed.
- * @retval -EBADMSG if the modem responded with an error to an AT command or the
- *         response could not be parsed.
- */
-int lte_lc_periodic_search_get(struct lte_lc_periodic_search_cfg *const cfg);
-
-/**
- * Clear the configured periodic search parameters.
- *
- * @retval 0 if the configuration was cleared.
- * @retval -EFAULT if an AT command could not be sent to the modem.
- * @retval -EBADMSG if the modem responded with an error to an AT command.
- */
-int lte_lc_periodic_search_clear(void);
-
-/**
- * Request an extra search.
- *
- * This can be used for example when modem is in sleep state between periodic searches. The search
- * is performed only when the modem is in sleep state between periodic searches.
- *
- * @retval 0 if the search request was successfully delivered to the modem.
- * @retval -EFAULT if an AT command could not be sent to the modem.
- */
-int lte_lc_periodic_search_request(void);
-
-/**
- * Read the current reduced mobility mode.
- *
- * @note This feature is only supported by modem firmware versions >= 1.3.2.
- *
- * @param[out] mode Reduced mobility mode.
- *
- * @retval 0 if a mode was found and written to the provided pointer.
- * @retval -EINVAL if input parameter was @c NULL.
- * @retval -EFAULT if an AT command failed.
- */
-int lte_lc_reduced_mobility_get(enum lte_lc_reduced_mobility_mode *mode);
-
-/**
- * Set reduced mobility mode.
- *
- * @note This feature is only supported by modem firmware versions >= 1.3.2.
- *
- * @param[in] mode Reduced mobility mode.
- *
- * @retval 0 if the new reduced mobility mode was accepted by the modem.
- * @retval -EFAULT if an AT command failed.
- */
-int lte_lc_reduced_mobility_set(enum lte_lc_reduced_mobility_mode mode);
-
-/**
- * Reset modem to factory settings.
- *
- * This operation is allowed only when the modem is not activated.
- *
- * @note This feature is only supported by modem firmware versions >= 1.3.0.
- *
- * @param[in] type Factory reset type.
- *
- * @retval 0 if factory reset was performed successfully.
- * @retval -EFAULT if an AT command failed.
- */
-int lte_lc_factory_reset(enum lte_lc_factory_reset_type type);
 
 /** @} */
