+++ lib/lte_link_control/CMakeLists.txt
@@ -5,10 +5,13 @@
 #
 
 zephyr_library()
+zephyr_library_include_directories(include)
 zephyr_library_sources(lte_lc.c)
-zephyr_library_sources(lte_lc_helpers.c)
 zephyr_library_sources(lte_lc_modem_hooks.c)
 zephyr_library_sources_ifdef(CONFIG_LTE_LC_TRACE lte_lc_trace.c)
 zephyr_library_sources_ifdef(CONFIG_LTE_SHELL lte_lc_shell.c)
+
+add_subdirectory(common)
+add_subdirectory(modules)
 
 zephyr_linker_sources(RODATA lte_lc.ld)
