+++ lib/lte_link_control/common/CMakeLists.txt
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2024 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+zephyr_library_sources(event_handler_list.c)
+zephyr_library_sources(work_q.c)
