+++ lib/lte_link_control/include/modules/cscon.h
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2024 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+#ifndef CSCON_H__
+#define CSCON_H__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int cscon_notifications_enable(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* CSCON_H__ */
