+++ lib/lte_link_control/include/modules/xmodemsleep.h
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2024 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+#ifndef XMODEMSLEEP_H__
+#define XMODEMSLEEP_H__
+
+#include <modem/lte_lc.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int xmodemsleep_notifications_enable(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* XMODEMSLEEP_H__ */
