+++ lib/lte_link_control/include/modules/xt3412.h
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2024 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+#ifndef XT3412_H__
+#define XT3412_H__
+
+#include <modem/lte_lc.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int xt3412_notifications_enable(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* XT3412_H__ */
