+++ samples/cellular/battery/prj.conf
@@ -19,6 +19,7 @@ CONFIG_NRF_MODEM_LIB=y
 
 # LTE link control
 CONFIG_LTE_LINK_CONTROL=y
+CONFIG_LTE_LC_CONEVAL=y
 
 # Stacks and heaps
 CONFIG_HEAP_MEM_POOL_SIZE=1024
