+++ samples/cellular/location/prj.conf
@@ -36,6 +36,7 @@ CONFIG_LTE_LINK_CONTROL=y
 CONFIG_LTE_EDRX_REQ=y
 # Request PSM active time of 8 seconds.
 CONFIG_LTE_PSM_REQ_RAT="00000100"
+CONFIG_LTE_LC_NCELLMEAS=y
 
 # AT Host library - Used to send AT commands directy from an UART terminal and to allow
 #		    integration with nRF Connect for Desktop LTE Link monitor application.
