+++ samples/cellular/modem_shell/prj.conf
@@ -119,6 +119,10 @@ CONFIG_LTE_LC_TAU_PRE_WARNING_THRESHOLD_MS=10240
 CONFIG_LTE_LC_MODEM_SLEEP_NOTIFICATIONS_THRESHOLD_MS=10240
 # Maximum number of neighbor cells for neighbor cell measurement
 CONFIG_LTE_NEIGHBOR_CELLS_MAX=17
+CONFIG_LTE_LC_REDMOB=y
+CONFIG_LTE_LC_PERIODICSEARCHCONF=y
+CONFIG_LTE_LC_NCELLMEAS=y
+CONFIG_LTE_LC_CONEVAL=y
 
 CONFIG_DATE_TIME=y
