+++ samples/cellular/nrf_cloud_multi_service/prj.conf
@@ -105,6 +105,7 @@ CONFIG_DATE_TIME=y
 
 # LTE link control - used by PGPS and main application
 CONFIG_LTE_LINK_CONTROL=y
+CONFIG_LTE_LC_NCELLMEAS=y
 
 # Settings - used by nRF Cloud library and PGPS
 CONFIG_SETTINGS=y
