+++ samples/cellular/nrf_cloud_rest_cell_location/prj.conf
@@ -33,6 +33,7 @@ CONFIG_CAF_BUTTONS_POLARITY_INVERSED=y
 
 # Modem/LTE Link
 CONFIG_LTE_LINK_CONTROL=y
+CONFIG_LTE_LC_NCELLMEAS=y
 
 # Modem info
 CONFIG_MODEM_INFO=y
