+++ tests/lib/location/prj.conf
@@ -29,3 +29,4 @@ CONFIG_MOCK_NRF_MODEM_AT=y
 CONFIG_LOG=n
 CONFIG_LOCATION_LOG_LEVEL_DBG=n
 CONFIG_LTE_LINK_CONTROL_LOG_LEVEL_DBG=n
+CONFIG_LTE_LC_NCELLMEAS=y
