+++ tests/lib/lte_lc_api/prj.conf
@@ -25,3 +25,13 @@ CONFIG_MOCK_NRF_MODEM_AT_SCANF_VARGS_STR_SIZE=32
 #CONFIG_LTE_LINK_CONTROL_LOG_LEVEL_DBG=y
 CONFIG_STACK_SENTINEL=y
 CONFIG_ENTROPY_GENERATOR=y
+
+CONFIG_LTE_LC_CONEVAL=y
+CONFIG_LTE_LC_EDRX=y
+CONFIG_LTE_LC_NCELLMEAS=y
+CONFIG_LTE_LC_PERIODICSEARCHCONF=y
+CONFIG_LTE_LC_PSM=y
+CONFIG_LTE_LC_REDMOB=y
+CONFIG_LTE_LC_XFACTORYRESET=y
+CONFIG_LTE_LC_XMODEMSLEEP=y
+CONFIG_LTE_LC_XT3412=y
